More UInt64 => IntPtr replacements
Former-commit-id: 4dd3b40b1fd474b54ecde68ee2e87a66a1c10106

+++ Anathema/Source/OS/Lua/LuaEngine.cs
@@ -16,7 +16,7 @@ namespace Anathema
             BindFunctions();
         }
 
-        public static String AddCodeInjectionTemplate(String Script, String ModuleName, UInt64 ModuleOffset)
+        public static String AddCodeInjectionTemplate(String Script, String ModuleName, IntPtr ModuleOffset)
         {
             String CodeInjection =
                 "-- No Description" + "\n\t\n" +
